Use BigDecimal instead of double

+++ src/main/kotlin/org/deblock/flights/controller/dto/FlightSearchResponseDTO.kt
@@ -1,5 +1,6 @@
 package org.deblock.flights.controller.dto
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class FlightSearchResponseDTO(val flights: List<FlightDTO>)
@@ -7,7 +8,7 @@ data class FlightSearchResponseDTO(val flights: List<FlightDTO>)
 data class FlightDTO(
     val airline: String,
     val supplier: String,
-    val fare: Double,
+    val fare: BigDecimal,
     val departureAirportCode: String,
     val destinationAirportCode: String,
     val departureDate: LocalDateTime,
