+++ src/main/kotlin/org/deblock/flights/service/Flight.kt
@@ -1,11 +1,12 @@
 package org.deblock.flights.service
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class Flight(
     val airline: String,
     val supplier: String,
-    val fare: Double,
+    val fare: BigDecimal,
     val departureAirportCode: String,
     val destinationAirportCode: String,
     val departureDate: Instant,
